// prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Product: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly price: CodecTypes['pg/numeric@1']['output'];
      readonly stockQty: CodecTypes['pg/int4@1']['output'];
      readonly active: CodecTypes['pg/bool@1']['output'];
      readonly categoryId: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
      readonly publicId: CodecTypes['pg/text@1']['output'] | null;
      readonly productId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Product: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly price: CodecTypes['pg/numeric@1']['input'];
      readonly stockQty: CodecTypes['pg/int4@1']['input'];
      readonly active: CodecTypes['pg/bool@1']['input'];
      readonly categoryId: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
      readonly publicId: CodecTypes['pg/text@1']['input'] | null;
      readonly productId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly category: {
      readonly createdAt: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly product: {
      readonly active: CodecTypes['pg/bool@1']['output'];
      readonly categoryId: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly price: CodecTypes['pg/numeric@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly stockQty: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly productImage: {
      readonly createdAt: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly productId: CodecTypes['pg/text@1']['output'];
      readonly publicId: CodecTypes['pg/text@1']['output'] | null;
      readonly url: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly category: {
      readonly createdAt: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly product: {
      readonly active: CodecTypes['pg/bool@1']['input'];
      readonly categoryId: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly price: CodecTypes['pg/numeric@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly stockQty: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly productImage: {
      readonly createdAt: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly productId: CodecTypes['pg/text@1']['input'];
      readonly publicId: CodecTypes['pg/text@1']['input'] | null;
      readonly url: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;